moved useBirdSearch.js
